requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-openai-api-bridge
google-generativeai>=0.3.1
pygame>=2.6.1
fastapi[all]
tiktoken
cloudscraper
google-genai
langgraph


# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain dependencies  
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai-api-bridge

# OpenAI client for testing
openai>=1.30.0

# Additional utilities
pydantic>=2.0.0
httpx>=0.25.0